feat: move project dependent elements to useProject
