+++ packages/nc-gui-v2/composables/useTheme/index.ts
@@ -1,9 +1,9 @@
 import { ConfigProvider } from 'ant-design-vue'
 import type { Theme as AntTheme } from 'ant-design-vue/es/config-provider'
 import tinycolor from 'tinycolor2'
-import { hexToRGB, themeV2Colors, useCssVar, useInjectionState, useProject } from '#imports'
+import { hexToRGB, themeV2Colors, useCssVar, useInjectionState } from '#imports'
 
-interface ThemeConfig extends AntTheme {
+export interface ThemeConfig extends AntTheme {
   primaryColor: string
   accentColor: string
 }
@@ -12,22 +12,10 @@ const [setup, use] = useInjectionState((config?: Partial<ThemeConfig>) => {
   const primaryColor = useCssVar('--color-primary', typeof document !== 'undefined' ? document.documentElement : null)
   const accentColor = useCssVar('--color-accent', typeof document !== 'undefined' ? document.documentElement : null)
 
-  const { project, projectMeta, updateProject } = useProject()
-
-  // set theme based on active project
-  watch(project, () => {
-    setTheme({
-      primaryColor: themeV2Colors['royal-blue'].DEFAULT,
-      accentColor: themeV2Colors.pink['500'],
-      ...projectMeta.value?.theme,
-    })
-  })
-
   /** current theme config */
   const currentTheme = ref({
     primaryColor: themeV2Colors['royal-blue'].DEFAULT,
     accentColor: themeV2Colors.pink['500'],
-    ...projectMeta.value?.theme,
   })
 
   /** set initial config */
@@ -35,8 +23,8 @@ const [setup, use] = useInjectionState((config?: Partial<ThemeConfig>) => {
 
   /** set theme (persists in localstorage) */
   function setTheme(theme: Partial<ThemeConfig>) {
-    const themePrimary = tinycolor(theme.primaryColor)
-    const themeAccent = tinycolor(theme.accentColor)
+    const themePrimary = theme?.primaryColor ? tinycolor(theme.primaryColor) : tinycolor(themeV2Colors['royal-blue'].DEFAULT)
+    const themeAccent = theme?.accentColor ? tinycolor(theme.accentColor) : tinycolor(themeV2Colors.pink['500'])
 
     // convert hex colors to rgb values
     primaryColor.value = themePrimary.isValid()
@@ -54,20 +42,9 @@ const [setup, use] = useInjectionState((config?: Partial<ThemeConfig>) => {
     })
   }
 
-  async function saveTheme(theme: Partial<ThemeConfig>) {
-    await updateProject({
-      meta: JSON.stringify({
-        ...projectMeta.value,
-        theme,
-      }),
-    })
-    setTheme(theme)
-  }
-
   return {
     theme: currentTheme,
     setTheme,
-    saveTheme,
   }
 }, 'theme')
